Sketcher: Group tests in different directories
